refactoring Quest & addCutScene init
