+++ Assets/Scenes/WorldScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5f6c2872f07e5c44593d0a644b96367e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
