+++ Assets/Scenes/WorldScene/MainCamera Profile.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c811aa7ca3f58849bef0f42e5992465
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
