--- Assets/Scripts/Logger/ForEditorBaseScripts/Quest.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Logger.ForEditorBaseScripts;
-using UnityEngine;
-
-namespace Ciart.Pagomoa.Logger.ForEditorBaseScripts
-{
-    [CreateAssetMenu(menuName = "MakeQuest")]
-    public class Quest : ScriptableObject
-    {
-        [SerializeField] public GameObject questInCharge;
-        [SerializeField] public string id;
-        [SerializeField] public List<string> prevQuestIds = new();
-        [SerializeField] public string title;
-        [TextArea, SerializeField] public string description;
-        
-        [SerializeField] public TextAsset startPrologue;
-        [SerializeField] public TextAsset completePrologue;
-        
-        [SerializeField] public Reward reward = new();
-        
-        public List<QuestCondition> questList;
-    }
-
-    [Serializable]
-    public class Reward
-    {
-        public int gold;
-        public ScriptableObject targetEntity;
-        public Sprite targetEntitySprite;
-        public int value;
-    }
-}
