--- Assets/Scripts/Logger/ForEditorBaseScripts/QuestCondition.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-namespace Logger.ForEditorBaseScripts
-{
-    [Serializable] 
-    public class QuestCondition
-    {
-        public QuestType questType;
-        
-        public string summary;
-        
-        public string value;
-
-        public ScriptableObject targetEntity;
-
-        public ConditionType conditionType;
-    }
-}
