--- Assets/Scripts/Logger/ForEditorBaseScripts/QuestCondition.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f7a992be34c24462b3e428040d869c68
-timeCreated: 1699271594
