+++ Assets/Scripts/Logger/ForEditorBaseScripts/QuestData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c10b6c981bc3f749956449d513366f9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
