--- Assets/Scripts/Logger/ProcessScripts/ProcessQuest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 29eecd23c0e93bd40bb9de5f301c70bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
