--- Assets/Scripts/Logger/ProcessScripts/ProgressedQuest.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 40a4e33485744b7e9bf40da465bb453c
-timeCreated: 1714287458
