+++ Assets/Scripts/Logger/ProcessScripts/QuestCondition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f82f123ecf4992c47bb5b2e989d61acb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
