+++ Assets/Scripts/Logger/QuestDatabase.cs
@@ -12,38 +12,16 @@ namespace Logger
 {
     public class QuestDatabase : MonoBehaviour
     {
-        public readonly List<ProgressedQuest> progressedQuests = new List<ProgressedQuest>();
-
         [Serializable]
         public class MapEntityQuest
         {
             public EntityController entity;
-            public List<Quest> entityQuests = new List<Quest>();
+            public List<QuestData> entityQuests = new List<QuestData>();
         }
 
         public List<MapEntityQuest> mapEntityQuests;
 
-        private ProgressedQuest FindQuestById(string id)
-        {
-            foreach (var quest in progressedQuests)
-            {
-                if (quest.id == id) return quest;
-            }
-            
-            return null;
-        }
-
-        public bool CheckQuestCompleteById(string id)
-        {
-            foreach (var quest in progressedQuests)
-            {
-                return FindQuestById(id) is not null && quest.accomplished;
-            }
-            
-            return false;
-        }
-
-        public Quest[] GetEntityQuestsByEntityID(EntityOrigin origin)
+        public QuestData[] GetEntityQuestsByEntityID(EntityOrigin origin)
         {
             foreach (var mapEntity in mapEntityQuests)
             {
@@ -53,7 +31,7 @@ namespace Logger
                 }
             }
 
-            return Array.Empty<Quest>();
+            return Array.Empty<QuestData>();
         }
     }
 }
