+++ Assets/Scripts/Logger/Quests/Quest1.asset
@@ -12,7 +12,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8c10b6c981bc3f749956449d513366f9, type: 3}
   m_Name: Quest1
   m_EditorClassIdentifier: 
-  questInCharge: {fileID: 0}
   id: copper_1
   prevQuestIds: []
   title: "\uAD6C\uB9AC\uB97C \uBAA8\uC544 \uC918!"
