+++ Assets/Scripts/Systems/Dialogue/EntityDialogue.cs
@@ -16,7 +16,7 @@ namespace Ciart.Pagomoa.Systems.Dialogue
 
         public Sprite portrait;
         private EntityController _entityController;
-        private Quest[] _entityQuests;
+        private QuestData[] _entityQuests;
 
         public Vector3 uiOffset = new Vector3(0f, 3.2f, 0f);
         private GameObject _questCompleteUI;
@@ -30,9 +30,9 @@ namespace Ciart.Pagomoa.Systems.Dialogue
             _questCompleteUI.transform.position += uiOffset;
         }
         
-        public Quest[] GetValidationQuests()
+        public QuestData[] GetValidationQuests()
         {
-           var result = new List<Quest>();
+           var result = new List<QuestData>();
 
            foreach (var quest in _entityQuests)
            {
@@ -110,7 +110,7 @@ namespace Ciart.Pagomoa.Systems.Dialogue
             
             _entityQuests = QuestManager.instance.database.GetEntityQuestsByEntityID(origin);
 
-            if (_entityQuests == Array.Empty<Quest>()) return; 
+            if (_entityQuests == Array.Empty<QuestData>()) return; 
             
             EventManager.AddListener<QuestCompleted>(OnCompleteQuestsUpdated);
         }
