+++ Assets/Timelines/ShopkeeperAppear.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff5aaa7558fea1841b0bfdcfdc96de55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
