+++ ProjectSettings/TagManager.asset
@@ -20,7 +20,7 @@ TagManager:
   - Light
   - Entity
   - Player
-  - 
+  - BackGround
   - CountUI
   - CutScene
   -
